Run init code before starting Gunicorn
