+++ Dockerfile
@@ -29,4 +29,5 @@ RUN mkdir -p /app/data
 RUN chmod -R 777 /app/data
 
 # Run app with gunicorn (production-grade WSGI)
+ENTRYPOINT [ "/app/entrypoint.sh" ]
 CMD ["gunicorn", "-w", "4", "-t", "240", "-b", "0.0.0.0:5050", "app:app"]
